requirements.txt
flask
flask_sqlalchemy
flask_marshmallow
pyjwt